Renamed train3d to train
