+++ configs/slurm_perlmutter_train_cm4.yaml
@@ -43,7 +43,7 @@ experiment:
   base_name: train
   sub_name: cm4_samudra
   rand_seed: 5
-  base_output_dir: train_3D
+  base_output_dir: train
   gantry: false
   cluster_data_dir: /pscratch/sd/s/suryad/data
   wandb:
