+++ scripts/.local-train
@@ -12,4 +12,4 @@ echo 'Using config file:' $CONFIG_PATH
 TMP_CONFIG_PATH=/tmp/config_temp.yaml
 sed "s|cluster_data_dir:.*|cluster_data_dir: $OCEAN_DATA_DIR|" $CONFIG_PATH > $TMP_CONFIG_PATH
 
-uv run torchrun -m src.ocean_emulators.train_3D --config $TMP_CONFIG_PATH --subname "test"
+uv run torchrun -m src.ocean_emulators.train --config $TMP_CONFIG_PATH --subname "test"
