+++ src/ocean_emulators/config.py
@@ -82,7 +82,7 @@ class ExperimentConfig:
     base_name: str = "train"
     sub_name: str = "cm4_samudra"
     rand_seed: int = 1
-    base_output_dir: str = "train_3D"
+    base_output_dir: str = "train"
     gantry: bool = False
     cluster_data_dir: str = "/"
     wandb: WandBConfig = field(default_factory=WandBConfig)
